feat(dashboard): replace static data with real metrics and improve UI interactions
The contribution grid now displays actual commit history with tooltips showing the date and commit count for each cell. Peak coding hours are grouped into time blocks (Morning, Afternoon, Evening, Night) instead of individual hours, with a dynamic summary indicating the most productive period. Repository cards are populated from the API using real repo data including language, status, branch, and relative last activity timestamps.

The logout button has been moved from the dashboard page into the header for consistent access across views. The API client handles 401 responses by redirecting to the login page. The server repos endpoint returns enriched data with computed health status and human-readable activity timestamps. The pull_requests migration schema uses BIGINT for github_id to support larger identifier values.

+++ server/src/db/migrate.ts
@@ -93,7 +93,7 @@ export async function migrate() {
       id              SERIAL PRIMARY KEY,
       user_id         INTEGER NOT NULL REFERENCES users(id) ON DELETE CASCADE,
       repo_id         INTEGER NOT NULL REFERENCES repos(id) ON DELETE CASCADE,
-      github_id       INTEGER NOT NULL,
+      github_id       BIGINT NOT NULL,
       number          INTEGER NOT NULL,
       title           TEXT NOT NULL,
       state           TEXT NOT NULL,
